backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
pydantic>=2.3.0
openai-whisper>=0.3.0
numpy<2.0.0,>=1.21.6
numba>=0.51.2,<0.57.0
llvmlite>=0.39.0
python-multipart>=0.0.6


# pip install -q -U google-genai

# # 更新pip
# pip install --upgrade pip setuptools wheel

# # 安装基础依赖
# pip install numpy==1.22.0
# pip install llvmlite==0.39.1
# pip install numba==0.56.4

# # 安装mecab
# brew install mecab

# # 设置编译环境变量
# export MACOSX_DEPLOYMENT_TARGET=11.0

# # 安装mecab-python
# pip install mecab-python3

# # 安装librosa (避免使用最新版本可能更稳定)
# pip install librosa==0.9.2

# # 最后安装TTS
# pip install TTS